# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: BCA Entertainment Works ERP API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /gst-invoices:
    get:
      operationId: getGstInvoices
      tags: [gst-invoices]
      summary: List GST invoices
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: month
          in: query
          schema:
            type: string
        - name: year
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: GST invoice list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GstInvoice"
    post:
      operationId: createGstInvoice
      tags: [gst-invoices]
      summary: Create GST invoice
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateGstInvoiceBody"
      responses:
        "201":
          description: Created invoice with line items
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GstInvoiceDetail"
...
  /gst-invoices/{id}:
    get:
      operationId: getGstInvoice
      tags: [gst-invoices]
      summary: Get GST invoice by id
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Invoice with line items and payments
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GstInvoiceDetail"
    patch:
      operationId: updateGstInvoice
      tags: [gst-invoices]
      summary: Update invoice notes or due date
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                notes:
                  type: string
                dueDate:
                  type: string
      responses:
        "200":
          description: Updated invoice
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GstInvoice"
  /gst-invoices/{id}/payments:
    post:
      operationId: recordInvoicePayment
      tags: [gst-invoices]
      summary: Record a payment against an invoice
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RecordPaymentBody"
      responses:
        "201":
          description: Payment recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvoicePayment"
  /gst-summary:
    get:
      operationId: getGstSummary
      tags: [gst-invoices]
      summary: Monthly GST summary — output tax, input tax, net liability
      parameters:
        - name: year
          in: query
          required: true
          schema:
            type: string
        - name: month
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: GST summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GstSummary"
  /supplier-bills:
    get:
      operationId: getSupplierBills
      tags: [gst-invoices]
      summary: List supplier bills
      parameters:
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Supplier bill list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SupplierBill"
    post:
      operationId: createSupplierBill
      tags: [gst-invoices]
      summary: Create supplier bill
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSupplierBillBody"
      responses:
        "201":
          description: Created bill
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SupplierBill"
...
  /supplier-bills/{id}:
    get:
      operationId: getSupplierBill
      tags: [gst-invoices]
      summary: Get supplier bill by id
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Supplier bill
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SupplierBill"
  /supplier-bills/{id}/pay:
    post:
      operationId: paySupplierBill
      tags: [gst-invoices]
      summary: Record payment against supplier bill
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [amount, paymentDate]
              properties:
                amount:
                  type: number
                paymentDate:
                  type: string
      responses:
        "200":
          description: Updated bill
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SupplierBill"
  /expenses:
    get:
      operationId: getExpenses
      tags: [gst-invoices]
      summary: List expenses
      parameters:
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Expense list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Expense"
    post:
      operationId: createExpense
      tags: [gst-invoices]
      summary: Create expense
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateExpenseBody"
      responses:
        "201":
          description: Created expense
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Expense"
...
  /expenses/{id}:
    patch:
      operationId: updateExpense
      tags: [gst-invoices]
      summary: Update expense
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateExpenseBody"
      responses:
        "200":
          description: Updated expense
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Expense"
    delete:
      operationId: deleteExpense
      tags: [gst-invoices]
      summary: Delete expense
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /expenses/{id}/approve:
    patch:
      operationId: approveExpense
      tags: [gst-invoices]
      summary: Approve expense
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Expense"
  /expenses/{id}/reject:
    patch:
      operationId: rejectExpense
      tags: [gst-invoices]
      summary: Reject expense
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Expense"
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
...
    GstInvoice:
      type: object
      properties:
        id:
          type: integer
        invoiceNumber:
          type: string
        invoiceDate:
          type: string
        customerName:
          type: string
        customerAddress:
          type: ["string", "null"]
        customerGstin:
          type: ["string", "null"]
        bcaGstin:
          type: ["string", "null"]
        workOrderId:
          type: ["integer", "null"]
        transactionType:
          type: string
        subtotal:
          type: string
        cgstAmount:
          type: string
        sgstAmount:
          type: string
        igstAmount:
          type: string
        roundOff:
          type: string
        total:
          type: string
        status:
          type: string
        paidAmount:
          type: string
        dueDate:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdById:
          type: ["integer", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - invoiceNumber
        - invoiceDate
        - customerName
        - transactionType
        - subtotal
        - cgstAmount
        - sgstAmount
        - igstAmount
        - roundOff
        - total
        - status
        - paidAmount
        - createdAt
        - updatedAt
...
    InvoicePayment:
      type: object
      properties:
        id:
          type: integer
        invoiceId:
          type: integer
        paymentDate:
          type: string
        amount:
          type: string
        paymentMode:
          type: string
        reference:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - invoiceId
        - paymentDate
        - amount
        - paymentMode
        - createdAt
    GstInvoiceDetail:
      allOf:
        - $ref: "#/components/schemas/GstInvoice"
        - type: object
          properties:
            lineItems:
              type: array
              items:
                $ref: "#/components/schemas/InvoiceLineItem"
            payments:
              type: array
              items:
                $ref: "#/components/schemas/InvoicePayment"
          required:
            - lineItems
            - payments
...
    CreateGstInvoiceBody:
      type: object
      required: [customerName, invoiceDate, lineItems]
      properties:
        customerName:
          type: string
        customerAddress:
          type: string
        customerGstin:
          type: string
        bcaGstin:
          type: string
        invoiceDate:
          type: string
        dueDate:
          type: string
        workOrderId:
          type: integer
        transactionType:
          type: string
        notes:
          type: string
        lineItems:
          type: array
          items:
            $ref: "#/components/schemas/CreateInvoiceLineItemBody"
    RecordPaymentBody:
      type: object
      required: [paymentDate, amount]
      properties:
        paymentDate:
          type: string
        amount:
          type: number
        paymentMode:
          type: string
        reference:
          type: string
        notes:
          type: string
...
    CreateSupplierBillBody:
      type: object
      required: [billNumber, supplierName, billDate, subtotal, gstAmount, total]
      properties:
        billNumber:
          type: string
        supplierName:
          type: string
        supplierGstin:
          type: string
        purchaseOrderId:
          type: integer
        billDate:
          type: string
        dueDate:
          type: string
        transactionType:
          type: string
          enum: [intrastate, interstate]
        subtotal:
          type: number
        gstAmount:
          type: number
        cgstAmount:
          type: number
        sgstAmount:
          type: number
        igstAmount:
          type: number
        total:
          type: number
        notes:
          type: string
...